requirements/base.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.4.0
PyJWT==2.10.1
django-cors-headers==4.6.0
django-filter==24.3
drf-spectacular==0.28.0
python-decouple==3.8
python-dateutil==2.9.0